requirements.txt
ddgs
requests
flask
python-dotenv